[Behat] Extract Payment transformer
