+++ etc/behat/services/contexts.xml
@@ -31,6 +31,7 @@
         <parameter key="sylius.behat.context.transform.lexical.class">Sylius\Behat\Context\Transform\LexicalContext</parameter>
         <parameter key="sylius.behat.context.transform.shared_storage.class">Sylius\Behat\Context\Transform\SharedStorageContext</parameter>
         <parameter key="sylius.behat.context.transform.channel.class">Sylius\Behat\Context\Transform\ChannelContext</parameter>
+        <parameter key="sylius.behat.context.transform.payment.class">Sylius\Behat\Context\Transform\PaymentContext</parameter>
 
         <parameter key="sylius.behat.context.hook.doctrine_orm.class">Sylius\Behat\Context\Hook\DoctrineORMContext</parameter>
 
@@ -173,6 +174,11 @@
 
         <service id="sylius.behat.context.transform.channel" class="%sylius.behat.context.transform.channel.class%" scope="scenario">
             <argument type="service" id="sylius.repository.channel" container="symfony" />
+            <tag name="sylius.behat.context" />
+        </service>
+
+        <service id="sylius.behat.context.transform.payment" class="%sylius.behat.context.transform.payment.class%" scope="scenario">
+            <argument type="service" id="sylius.repository.payment_method" container="symfony" />
             <tag name="sylius.behat.context" />
         </service>
