+++ etc/behat/suites/domain_order.yml
@@ -16,6 +16,7 @@ default:
                 - sylius.behat.context.transform.shared_storage
                 - sylius.behat.context.transform.addressing
                 - sylius.behat.context.transform.customer
+                - sylius.behat.context.transform.payment
 
                 - sylius.behat.context.domain.order
