+++ src/Sylius/Behat/Context/Setup/PaymentContext.php
@@ -78,17 +78,4 @@ final class PaymentContext implements Context
 
         $this->paymentMethodRepository->add($paymentMethod);
     }
-
-    /**
-     * @Transform /^"([^"]+)" payment$/
-     */
-    public function getPaymentMethodByName($paymentMethodName)
-    {
-        $paymentMethod = $this->paymentMethodRepository->findOneBy(['name' => $paymentMethodName]);
-        if (null === $paymentMethod) {
-            throw new \InvalidArgumentException(sprintf('Payment method with name "%s" does not exist.', $paymentMethodName));
-        }
-
-        return $paymentMethod;
-    }
 }
